profiles/postgresql-qdrant-redis/requirements.txt
python-dotenv==0.21.0
# llama-index
openai
langchain
langchainhub
langchain-community
langchain-openai
streamlit
fastapi
requests
sse-starlette==2.0.0
langchain-qdrant
emtechstack